backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
